requirement.txt
requests  
beautifulsoup4  
xlwt  
fake_useragent  
selenium  
undetected_chromedriver  
webdriver-manager  
